Added loading of Stations back from DB in ConsoleApp for Td3200 import

+++ DataSetExtension/Station.cs
@@ -4,7 +4,7 @@ namespace DataSetExtension
 {
     public class Station
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>
         /// This is the Station Id (natural key)
@@ -13,52 +13,52 @@ namespace DataSetExtension
 
         public string Name { get; set; }
 
-        public int GridPoint { get; set; }
+        public long GridPoint { get; set; }
 
         /// <summary>
         /// Denotes the sequence of this station in relation to GridPoint (0 = Primary, 1 = secondary, 2...n)
         /// </summary>
-        public int Sequence { get; set; }
+        public long Sequence { get; set; }
 
         /// <summary>
         /// Latitude of the Station
         /// </summary>
-        public int Latitude { get; set; }
+        public long Latitude { get; set; }
 
         /// <summary>
         /// Longitude of the Station
         /// </summary>
-        public int Longitude { get; set; }
+        public long Longitude { get; set; }
 
         /// <summary>
         /// Rounded Latitude of the Grid Point
         /// </summary>
-        public int GridPointLatitude { get; set; }
+        public long GridPointLatitude { get; set; }
 
         /// <summary>
         /// Rounded Longitude of the Grid Point
         /// </summary>
-        public int GridPointLongitude { get; set; }
+        public long GridPointLongitude { get; set; }
 
-        public int HistoricalRecordCount { get; set; }
+        public long HistoricalRecordCount { get; set; }
 
-        public int RecordCount { get; set; }
+        public long RecordCount { get; set; }
 
         public virtual void Parse(string record)
         {
             var gridpoint = record.Substring(0, 3).Trim();
             if (gridpoint != "")
             {
-                GridPoint = int.Parse(record.Substring(0, 3).Trim());
-                GridPointLatitude = int.Parse(record.Substring(4, 3).Trim());
-                GridPointLongitude = int.Parse(record.Substring(7, 3));
+                GridPoint = long.Parse(record.Substring(0, 3).Trim());
+                GridPointLatitude = long.Parse(record.Substring(4, 3).Trim());
+                GridPointLongitude = long.Parse(record.Substring(7, 3));
             }
 
             Name = record.Substring(10, 30).Trim();
             Number = record.Substring(41, 7).Trim();
-            Latitude = int.Parse(record.Substring(48, 5).Trim());
-            Longitude = int.Parse(record.Substring(53, 6).Trim());
-            HistoricalRecordCount = int.Parse(record.Substring(59, 8).Trim());
+            Latitude = long.Parse(record.Substring(48, 5).Trim());
+            Longitude = long.Parse(record.Substring(53, 6).Trim());
+            HistoricalRecordCount = long.Parse(record.Substring(59, 8).Trim());
         }
 
         public virtual void Parse(string record, Station parent)
